+++ domain/entity/order.go
@@ -10,7 +10,7 @@ type Order struct {
 	OrderId        string                `json:"orderId"`
 	UserId         string                `json:"userId"`
 	Cart           []valueObject.Product `json:"cart"`
-	TotalPrice     float64               `json:"totalPrice"`
+	TotalPrice     int64                 `json:"totalPrice"`
 	Status         string                `json:"status"`
 	ShippingMethod string                `json:"shippingMethod"`
 	OrderDate      time.Time             `json:"orderDate"`
